+++ async_pokepy/http.py
@@ -37,25 +37,31 @@ from .utils import _fmt_param
 LOG = logging.getLogger(__name__)
 
 
+class Route:
+    __slots__ = ("params", "url")
+
+    def __init__(self, base, path, *args):
+        self.params = [_fmt_param(arg) for arg in args]
+
+        self.url = (base + path + "/".join(self.params))
+
+
 class HTTPPokemonClient:
-    __slots__ = ("loop", "headers", "_session", "_lock")
+    __slots__ = ("loop", "headers", "_session", "_lock", "base")
 
-    BASE = "https://pokeapi.co/api/v2"
-
-    def __init__(self, loop: asyncio.AbstractEventLoop = None):
+    def __init__(self, loop: asyncio.AbstractEventLoop = None, base: str = None, user_agent: str = None):
         self.loop = loop or asyncio.get_event_loop()
         self.headers = {
-            "User-Agent": "Python/{0[0]}.{0[1]} aiohttp/{1}".format(sys.version_info, aiohttp.__version__)
+            "User-Agent": user_agent or "Python/{0[0]}.{0[1]} aiohttp/{1}".format(sys.version_info, aiohttp.__version__)
         }
         self._session = None
         self._lock = asyncio.Lock(loop=self.loop)
+        self.base = base or "https://pokeapi.co/api/v2"
 
-    async def request(self, url: str, **params) -> Union[str, dict]:
-        url = (self.BASE + url)
-
+    async def request(self, route, **params) -> Union[str, dict]:
         async with self._lock:
             for tries in range(5):
-                async with self._session.get(url, **params) as resp:
+                async with self._session.get(route.url, **params) as resp:
                     LOG.info("%s %s returned %d status code", resp.method, resp.url, resp.status)
 
                     data = (await resp.json() if "application/json" in resp.headers["Content-Type"]
@@ -107,10 +113,10 @@ class HTTPPokemonClient:
             raise PokeAPIException(resp, "Failed to get sprite.")
 
     def get_pokemon(self, query: Union[int, str]) -> Coroutine:
-        return self.request("/pokemon/{0}".format(_fmt_param(query)))
+        return self.request(Route(self.base, "/pokemon/", query))
 
     def get_move(self, query: Union[int, str]) -> Coroutine:
-        return self.request("/move/{0}".format(_fmt_param(query)))
+        return self.request(Route(self.base, "/move/", query))
 
     def get_ability(self, query: Union[int, str]) -> Coroutine:
-        return self.request("/ability/{0}".format(_fmt_param(query)))
+        return self.request(Route(self.base, "/ability/", query))