+++ async_pokepy/types/other.py
@@ -24,6 +24,8 @@ FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 DEALINGS IN THE SOFTWARE.
 """
 
+from ..utils import _pretty_format
+
 
 class VersionGameIndex:
     """Represents a partial Version object bound to a GameIndex.
@@ -36,7 +38,7 @@ class VersionGameIndex:
         The name of the version relevant to this game index.`"""
     def __init__(self, data: dict):
         self.game_index = data["game_index"]
-        self.version = data["version"]["name"]
+        self.version = _pretty_format(data["version"]["name"])
 
     def __repr__(self) -> str:
         return "<VersionGameIndex game_index={0.game_index} version='{0.version}'>".format(self)