+++ docs/changelog.rst
@@ -6,6 +6,17 @@ Changelog
 This page aims to keep an organized view of all of the changes
 since the wrapper's initial release.
 
+0.0.8a
+------
+
+- :meth:`Client.read_sprite` and :meth:`Client.save_sprite`
+  for easy sprite downloading.
+- You can now customize the base uri to make requests to
+  and the User-Agent header.
+- Pretty formatted names, e.g ``thick-fat`` is now ``Thick Fat``.
+- You can now search full objects by both id and name.
+- Improvements in internal routing.
+
 0.0.7a
 ------